% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attrition_rate_df.R
\name{get_attrition_rate_df}
\alias{get_attrition_rate_df}
\title{Get attrition rate dataframe}
\usage{
get_attrition_rate_df(data, entry_year, tenure, attrition_flg)
}
\arguments{
\item{data}{dataframe}

\item{entry_year}{string}

\item{tenure}{integer}

\item{attrition_flg}{interger}
}
\value{
dataframe
}
\description{
Get attrition rate dataframe
}
\examples{
set.seed(123)
df <- data.frame(
  entry_year = rep(c("FY17", "FY18", "FY19"), 3),
  tenure = c(0, 0, 0, 1, 1, 0, 2, 1, 0),
  attrition_flg = sample(c(0, 1), 9, replace = TRUE),
  stringsAsFactors = FALSE
)
get_attrition_rate_df(df, entry_year, tenure, attrition_flg)
}
