% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hankana_to_zenkana.R
\name{hankana_to_zenkana}
\alias{hankana_to_zenkana}
\title{Replace hankakukatakana with zenkakukatakana}
\usage{
hankana_to_zenkana(string)
}
\arguments{
\item{string}{}
}
\value{
string
}
\description{
Replace hankakukatakana with zenkakukatakana
}
\examples{
string <- "ﾊﾝｶｸｶﾅﾊｱﾌﾞﾅｲ"
hankana_to_zenkana(string)
}
