% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_to_int.R
\name{sign_to_int}
\alias{sign_to_int}
\title{Replace sign with integer}
\usage{
sign_to_int(sign)
}
\arguments{
\item{sign}{string}
}
\value{
integer
}
\description{
Replace sign with integer
}
\examples{
sign <- c("A+", "B", NA_character_, "N", "-", "B+", "D")
sign_to_int(sign)
}
